+++ src/main/java/com/kosa/realestate/InvalidPasswordException.java
@@ -1,4 +1,4 @@
-package com.kosa.realestate.users;
+package com.kosa.realestate;
 
 public class InvalidPasswordException extends RuntimeException {
   public InvalidPasswordException(String message) {
