+++ src/main/java/com/kosa/realestate/security/UserRole.java
@@ -1,4 +1,4 @@
-package com.kosa.realestate.users;
+package com.kosa.realestate.security;
 
 import lombok.Getter;
